get_cards file and fixed straight function
